+++ apps/api_admin/__init__.py
@@ -13,10 +13,9 @@ from .resources import order
 from .resources import question
 from .resources import token
 from .resources import user
-from .resources import test
 from .resources import code
 
-v1_admin_app: FastAPI = FastAPI(include_in_schema=Config.INCLUDE_IN_SCHEMA)
+api_admin_app: FastAPI = FastAPI(include_in_schema=Config.INCLUDE_IN_SCHEMA)
 
 
 def register_routers(app: FastAPI):
@@ -28,15 +27,14 @@ def register_routers(app: FastAPI):
     app.include_router(question.router, prefix='/questions', tags=['questions'])
     app.include_router(token.router, prefix='/tokens', tags=['tokens'])
     app.include_router(user.router, prefix='/users', tags=['users'])
-    app.include_router(test.router, prefix='/tests', tags=['tests'])
     app.include_router(code.router, prefix='/code', tags=['code'])
 
 
 def init_sub_app(app: FastAPI):
     """mount sub app"""
 
-    register_exception(v1_admin_app)
-    register_routers(v1_admin_app)
-    app.mount(path='/api/v1/admin', app=v1_admin_app, name='v1_admin_app')
+    register_exception(api_admin_app)
+    register_routers(api_admin_app)
+    app.mount(path='/api/admin', app=api_admin_app, name='api_admin_app')
 
     return app