+++ apps/api_admin/resources/file.py
@@ -3,7 +3,7 @@ from typing import Optional, List
 from fastapi import APIRouter, UploadFile, File
 
 from extensions import Route, error_response, resp_success
-from apps.v1_admin.entities import FileSchema
+from apps.api_admin.entities import FileSchema
 
 router = APIRouter(route_class=Route, responses=error_response)