+++ sockets/__init__.py
@@ -1,0 +1,1 @@
+from .server import socket_io, init_sub_app